warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@11251 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/IcqOscarJ/src/fam_01service.cpp
@@ -393,7 +393,7 @@ void CIcqProto::handleServiceFam(BYTE *pBuffer, size_t wBufferLength, snac_heade
 		while (wBufferLength > 4) { // loop thru all items
 			WORD itemType = pBuffer[0] * 0x10 | pBuffer[1];
 			BYTE itemFlags = pBuffer[2];
-			BYTE itemLen = pBuffer[3];
+			size_t itemLen = pBuffer[3];
 
 			if (itemType == AVATAR_HASH_PHOTO) { /// TODO: handle photo item
 				// skip photo item
